turtlegifmaker/requirements.txt
invoke==1.4.1
Pillow==8.0.1
